requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# RAG — FAISS + sentence embeddings (replaces chromadb)
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0

# Document parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0

python-dotenv>=1.0.0
